fix: prr-56 move Hero component to layout helper components
